requirement.txt
asgiref==3.5.2
boto3==1.23.5
botocore==1.26.5
certifi==2022.5.18.1
charset-normalizer==2.0.12
Django==4.0.4
django-cors-headers==3.12.0
django-smuggler==1.0.4
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
idna==3.3
jmespath==1.0.0
Pillow==9.1.1
PyJWT==2.4.0
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
